decouple AbstractChangeParentClassRector
